+++ protected/modules/articles/migrations/m000102_000001_create_page_articles.php
@@ -29,7 +29,7 @@ class m000102_000001_create_page_articles extends Migration
         $this->createTable($this->tableName, [
             'id' => $this->primaryKey(),
 
-            'published_at' => $this->integer(),
+            'published_at' => $this->bigInteger(),
             'status' => $this->smallInteger()->notNull()->defaultValue(Articles::STATUS_DRAFT),
 
             'content_title' => $this->string()->notNull(),
