+++ protected/modules/news/migrations/m000103_000001_create_page_news.php
@@ -29,7 +29,7 @@ class m000103_000001_create_page_news extends Migration
         $this->createTable($this->tableName, [
             'id' => $this->primaryKey(),
 
-            'published_at' => $this->integer(),
+            'published_at' => $this->bigInteger(),
             'status' => $this->smallInteger()->notNull()->defaultValue(News::STATUS_DRAFT),
 
             'content_title' => $this->string()->notNull(),
